--- src/components/Pages/Tracker/index.ts
@@ -1,1 +1,0 @@
-export * from './Tracker'
