+++ src/components/Pages/index.ts
@@ -1,9 +1,7 @@
-import { Tracker } from './Tracker'
 import { ConfPage } from './ConfPage'
 import { FrontPage } from './FrontPage'
 
 export {
-  Tracker,
   ConfPage,
   FrontPage
 }
