+++ src/components/index.ts
@@ -5,6 +5,7 @@ import { ConfPageHeader } from 'components/ConfPageHeader'
 import { ResultDetails } from 'components/ResultDetails'
 import { SearchInput } from 'components/SearchInput'
 import { ScrollToTop } from 'components/HOCS/ScrollToTop'
+import { Tracker } from 'components/HOCS/Tracker'
 import { Header } from 'components/Header'
 import { Video } from 'components/Video'
 import { Meta } from 'components/Meta'
@@ -17,6 +18,7 @@ export {
   Meta,
   Video,
   Header,
+  Tracker,
   ConfPage,
   FrontPage,
   ScrollToTop,
