+++ src/redux/modules/search.ts
@@ -31,7 +31,6 @@ export const textInDetails = (filterValue: string, termsToSearch: [string, strin
   any((phrase) => phrase.includes(filterValue), termsToSearch)
 
 // filters videos on a conference
-//
 export const filterVideos = (
   videos: IndexedVideos,
   presenters: IndexedPresenters,
@@ -84,8 +83,8 @@ const computeFilteredConferences = ({
     const matchedVideos =
       // if we think user is searching for a conference, returns all it's videos
       cleanQuery(title).includes(filterValue) ? conferenceVideos :
-      // otherwise just filter them
-      filterVideos(videos, presenters, conferences, filterValue, conferenceKey)
+        // otherwise just filter them
+        filterVideos(videos, presenters, conferences, filterValue, conferenceKey)
 
     // return conference if it has any matched videos
     return createConference(
@@ -102,37 +101,37 @@ const computeFilteredConferences = ({
 export const filterEpic: Epic<any, any, ApplicationState> = (action$, store) =>
   action$
     .ofType(FILTER).pipe(
-    debounceTime(80),
-    map((action) => {
-      const { payload: filterValue = ''} = action
-      const {
-        data: {
-          conferences, videosSearchable, presentersSearchable
-        },
-        conferencePage: { selectedConferenceId },
-        router
-      } = store.value
-      const rAction: Action<IndexedConferences> = { type: SET_FILTERED_CONFERENCES }
-      let filteredConferences = conferences;
+      debounceTime(80),
+      map((action) => {
+        const { payload: filterValue = '' } = action
+        const {
+          data: {
+            conferences, videosSearchable, presentersSearchable
+          },
+          conferencePage: { selectedConferenceId },
+          router
+        } = store.value
+        const rAction: Action<IndexedConferences> = { type: SET_FILTERED_CONFERENCES }
+        let filteredConferences = conferences;
 
-      // if on a conference page, only filter videos of that conference
-      if (router.location.pathname.includes('conference')) {
-        filteredConferences = {
-          [selectedConferenceId]: conferences[selectedConferenceId]
+        // if on a conference page, only filter videos of that conference
+        if (router.location.pathname.includes('conference')) {
+          filteredConferences = {
+            [selectedConferenceId]: conferences[selectedConferenceId]
+          }
         }
-      }
 
-      // if no/empty query, return original/all set of videos
-      rAction.payload = isFilterEmpty(filterValue) ?
-        filteredConferences :
-        computeFilteredConferences({
-          filterValue: cleanQuery(filterValue),
-          conferences: filteredConferences,
-          presenters: presentersSearchable,
-          videos: videosSearchable
-        })
-      return rAction
-    }), )
+        // if no/empty query, return original/all set of videos
+        rAction.payload = isFilterEmpty(filterValue) ?
+          filteredConferences :
+          computeFilteredConferences({
+            filterValue: cleanQuery(filterValue),
+            conferences: filteredConferences,
+            presenters: presentersSearchable,
+            videos: videosSearchable
+          })
+        return rAction
+      }))
 
 export const searchEpics = combineEpics(filterEpic)
 export const searchActions = {
